+++ CHANGELOG.md
@@ -37,6 +37,7 @@ Compatibility:
 * Make `Kernel#{caller, caller_locations}` and `Thread#backtrace_locations` compatible with endless ranges.
 * Implemented `Dir#each_child`.
 * Implemented `Kernel.{chomp, chop}` and `Kernel#{chomp, chop}`.
+* Implemented `-p` and `-a`, and `-l` CLI options.
 
 Performance: